+++ src/ansys/additive/core/parametric_study/parametric_runner.py
@@ -50,6 +50,7 @@ class ParametricRunner:
     def simulate(
         df: pd.DataFrame,
         additive: Additive,
+        simulation_ids: list[str] = None,
         type: list[SimulationType] = None,
         priority: int = None,
         iteration: int = None,
@@ -65,6 +66,10 @@ class ParametricRunner:
             Parametric study data frame.
         additive : Additive
             Additive service connection to use for running simulations.
+        simulation_ids: list[str], default: None
+            List of simulation IDs to run. The default is ``None``, in which case
+            all simulations in the parametric study are run. Any other filtering criteria
+            if provided, are applied on this list.
         type : list, default: None
             List of the simulation types to run. The default is ``None``, in which case all
             simulation types are run.
@@ -92,6 +97,18 @@ class ParametricRunner:
         view = df[
             (df[ColumnNames.STATUS] == SimulationStatus.PENDING) & df[ColumnNames.TYPE].isin(type)
         ]
+
+        if isinstance(simulation_ids, list) and len(simulation_ids) > 0:
+            simulation_ids_list = list()
+            for sim_id in simulation_ids:
+                if sim_id not in view[ColumnNames.ID].values:
+                    LOG.warning(f"Simulation ID '{sim_id}' not found in the parametric study")
+                elif sim_id in simulation_ids_list:
+                    LOG.debug(f"Simulation ID '{sim_id}' has already been added")
+                else:
+                    simulation_ids_list.append(sim_id)
+            view = view[view[ColumnNames.ID].isin(simulation_ids_list)]
+
         if priority is not None:
             view = view[view[ColumnNames.PRIORITY] == priority]
         view = view.sort_values(by=ColumnNames.PRIORITY, ascending=True)
@@ -168,25 +185,29 @@ class ParametricRunner:
     def _create_single_bead_input(
         row: pd.Series, material: AdditiveMaterial, machine: AdditiveMachine
     ) -> SingleBeadInput:
-        return SingleBeadInput(
-            id=row[ColumnNames.ID],
+        sb_input = SingleBeadInput(
             material=material,
             machine=machine,
             bead_length=row[ColumnNames.SINGLE_BEAD_LENGTH],
         )
+        # overwrite the ID value with the simulation ID from the table
+        sb_input._id = row[ColumnNames.ID]
+        return sb_input
 
     @staticmethod
     def _create_porosity_input(
         row: pd.Series, material: AdditiveMaterial, machine: AdditiveMachine
     ) -> PorosityInput:
-        return PorosityInput(
-            id=row[ColumnNames.ID],
+        p_input = PorosityInput(
             material=material,
             machine=machine,
             size_x=row[ColumnNames.POROSITY_SIZE_X],
             size_y=row[ColumnNames.POROSITY_SIZE_Y],
             size_z=row[ColumnNames.POROSITY_SIZE_Z],
         )
+        # overwrite the ID value with the simulation ID from the table
+        p_input._id = row[ColumnNames.ID]
+        return p_input
 
     @staticmethod
     def _create_microstructure_input(
@@ -199,8 +220,7 @@ class ParametricRunner:
             or not np.isnan(row[ColumnNames.MICRO_MELT_POOL_DEPTH])
         )
 
-        return MicrostructureInput(
-            id=row[ColumnNames.ID],
+        ms_input = MicrostructureInput(
             material=material,
             machine=machine,
             sample_size_x=row[ColumnNames.MICRO_SIZE_X],
@@ -249,3 +269,6 @@ class ParametricRunner:
                 else MicrostructureInput.DEFAULT_RANDOM_SEED
             ),
         )
+        # overwrite the ID value with the simulation ID from the table
+        ms_input._id = row[ColumnNames.ID]
+        return ms_input