+++ src/ansys/additive/core/parametric_study/parametric_study.py
@@ -32,7 +32,6 @@ import warnings
 
 import dill
 import numpy as np
-import pandas as pd
 
 from ansys.additive.core import (
     LOG,
@@ -57,11 +56,12 @@ from .constants import DEFAULT_ITERATION, DEFAULT_PRIORITY, FORMAT_VERSION, Colu
 from .parametric_runner import ParametricRunner
 from .parametric_utils import build_rate, energy_density
 
-# Suppress: FutureWarning: The behavior of DataFrame concatenation with empty or
+# Suppress: FutureWarning in pandas: The behavior of DataFrame concatenation with empty or
 # all-NA entries is deprecated. In a future version, this will no longer exclude
 # empty or all-NA columns when determining the result dtypes. To retain the old
 # behavior, exclude the relevant entries before the concat operation.
 warnings.simplefilter(action="ignore", category=FutureWarning)
+import pandas as pd  # noqa: E402
 
 
 def save_on_return(func):
@@ -175,6 +175,7 @@ class ParametricStudy:
     def run_simulations(
         self,
         additive: Additive,
+        simulation_ids: list[str] | None = None,
         type: list[SimulationType] | None = None,
         priority: int | None = None,
         iteration: int = None,
@@ -189,6 +190,9 @@ class ParametricStudy:
         ----------
         additive : Additive
             Additive service connection to use for running simulations.
+        simulation_ids : list[str], default: None
+            List of simulation IDs to run. If this value is ``None``,
+            all simulations with a status of ``Pending`` are run.
         type : list[SimulationType], default: None
             Type of simulations to run. If this value is ``None``,
             all simulation types are run.
@@ -202,6 +206,7 @@ class ParametricStudy:
         summaries = ParametricRunner.simulate(
             self.data_frame(),
             additive,
+            simulation_ids=simulation_ids,
             type=type,
             priority=priority,
             iteration=iteration,
@@ -1341,7 +1346,7 @@ class ParametricStudy:
 
             dict[ColumnNames.ITERATION] = iteration
             dict[ColumnNames.PRIORITY] = priority
-            dict[ColumnNames.ID] = self._create_unique_id(id=input.id)
+            dict[ColumnNames.ID] = input.id
             dict[ColumnNames.STATUS] = status
             dict[ColumnNames.MATERIAL] = input.material.name
             dict[ColumnNames.LASER_POWER] = input.machine.laser_power
@@ -1579,9 +1584,9 @@ class ParametricStudy:
         if id is not None and not self._data_frame[ColumnNames.ID].str.match(f"{id}").any():
             return id
         _prefix = id or prefix or "sim"
-        uid = f"{_prefix}_{misc.short_uuid()}"
+        uid = f"{_prefix}_{misc.short_uuid(6)}"
         while self._data_frame[ColumnNames.ID].str.match(f"{uid}").any():
-            uid = f"{_prefix}_{misc.short_uuid()}"
+            uid = f"{_prefix}_{misc.short_uuid(6)}"
         return uid
 
     @save_on_return