+++ tests/parametric_study/test_parametric_runner.py
@@ -39,6 +39,8 @@ from ansys.additive.core.additive import Additive
 import ansys.additive.core.parametric_study as ps
 from ansys.additive.core.parametric_study.constants import ColumnNames
 from ansys.additive.core.parametric_study.parametric_runner import ParametricRunner as pr
+from ansys.additive.core.single_bead import SingleBeadSummary
+from tests import test_utils
 
 
 def test_create_machine_assigns_all_values():
@@ -121,11 +123,10 @@ def test_create_machine_assigns_default_values():
 
 def test_create_single_bead_input():
     # arrange
-    id = "test_id"
     bead_length = 9.5e-3
     series = pd.Series(
         {
-            ColumnNames.ID: id,
+            ColumnNames.ID: "test_id",
             ColumnNames.SINGLE_BEAD_LENGTH: bead_length,
         }
     )
@@ -137,7 +138,7 @@ def test_create_single_bead_input():
 
     # assert
     assert isinstance(input, SingleBeadInput)
-    assert input.id == id
+    assert input.id == "test_id"
     assert input.bead_length == bead_length
     assert input.machine == machine
     assert input.material == material
@@ -285,9 +286,9 @@ def test_simulate_sorts_by_priority(tmp_path: pytest.TempPathFactory):
     # arrange
     study = ps.ParametricStudy(tmp_path / "test_study")
     material = AdditiveMaterial(name="test_material")
-    sb = SingleBeadInput(id="test_1", material=material)
-    p = PorosityInput(id="test_2", material=material)
-    ms = MicrostructureInput(id="test_3", material=material)
+    sb = SingleBeadInput(material=material)
+    p = PorosityInput(material=material)
+    ms = MicrostructureInput(material=material)
     study.add_inputs([sb], priority=1)
     study.add_inputs([p], priority=2)
     study.add_inputs([ms], priority=3)
@@ -306,9 +307,9 @@ def test_simulate_filters_by_priority(tmp_path: pytest.TempPathFactory):
     # arrange
     study = ps.ParametricStudy(tmp_path / "test_study")
     material = AdditiveMaterial(name="test_material")
-    sb = SingleBeadInput(id="test_1", material=material)
-    p = PorosityInput(id="test_2", material=material)
-    ms = MicrostructureInput(id="test_3", material=material)
+    sb = SingleBeadInput(material=material)
+    p = PorosityInput(material=material)
+    ms = MicrostructureInput(material=material)
     study.add_inputs([sb], priority=1)
     study.add_inputs([p], priority=2)
     study.add_inputs([ms], priority=3)
@@ -327,9 +328,9 @@ def test_simulate_filters_by_iteration(tmp_path: pytest.TempPathFactory):
     # arrange
     study = ps.ParametricStudy(tmp_path / "test_study")
     material = AdditiveMaterial(name="test_material")
-    sb1 = SingleBeadInput(id="iteration_1", material=material, bead_length=0.001)
-    sb2 = SingleBeadInput(id="iteration_2", material=material, bead_length=0.002)
-    sb3 = SingleBeadInput(id="iteration_3", material=material, bead_length=0.003)
+    sb1 = SingleBeadInput(material=material, bead_length=0.001)
+    sb2 = SingleBeadInput(material=material, bead_length=0.002)
+    sb3 = SingleBeadInput(material=material, bead_length=0.003)
     study.add_inputs([sb1], iteration=1)
     study.add_inputs([sb2], iteration=2)
     study.add_inputs([sb3], iteration=3)
@@ -348,9 +349,9 @@ def test_simulate_filters_by_single_simulation_type(tmp_path: pytest.TempPathFac
     # arrange
     study = ps.ParametricStudy(tmp_path / "test_study")
     material = AdditiveMaterial(name="test_material")
-    sb = SingleBeadInput(id="test_1", material=material)
-    p = PorosityInput(id="test_2", material=material)
-    ms = MicrostructureInput(id="test_3", material=material)
+    sb = SingleBeadInput(material=material)
+    p = PorosityInput(material=material)
+    ms = MicrostructureInput(material=material)
     study.add_inputs([sb], priority=1)
     study.add_inputs([p], priority=2)
     study.add_inputs([ms], priority=3)
@@ -369,9 +370,9 @@ def test_simulate_filters_by_simulation_type_list(tmp_path: pytest.TempPathFacto
     # arrange
     study = ps.ParametricStudy(tmp_path / "test_study")
     material = AdditiveMaterial(name="test_material")
-    sb = SingleBeadInput(id="test_1", material=material)
-    p = PorosityInput(id="test_2", material=material)
-    ms = MicrostructureInput(id="test_3", material=material)
+    sb = SingleBeadInput(material=material)
+    p = PorosityInput(material=material)
+    ms = MicrostructureInput(material=material)
     study.add_inputs([sb], priority=1)
     study.add_inputs([p], priority=2)
     study.add_inputs([ms], priority=3)
@@ -394,9 +395,9 @@ def test_simulate_skips_simulations_with_missing_materials(tmp_path: pytest.Temp
     # arrange
     study = ps.ParametricStudy(tmp_path / "test_study")
     material = AdditiveMaterial(name="test_material")
-    sb = SingleBeadInput(id="test_1", material=material)
-    p = PorosityInput(id="test_2", material=material)
-    ms = MicrostructureInput(id="test_3", material=material)
+    sb = SingleBeadInput(material=material)
+    p = PorosityInput(material=material)
+    ms = MicrostructureInput(material=material)
     study.add_inputs([sb], priority=1)
     study.add_inputs([p], priority=2)
     study.add_inputs([ms], priority=3)
@@ -417,7 +418,7 @@ def test_simulate_returns_empty_list_when_no_simulations_meet_criteria(
     # arrange
     study = ps.ParametricStudy(tmp_path / "test_study")
     material = AdditiveMaterial(name="test_material")
-    sb = SingleBeadInput(id="test_1", material=material)
+    sb = SingleBeadInput(material=material)
     study.add_inputs([sb], priority=1)
     mock_additive = create_autospec(Additive)
     caplog.set_level(logging.WARNING, logger="PyAdditive_global")
@@ -431,3 +432,214 @@ def test_simulate_returns_empty_list_when_no_simulations_meet_criteria(
     for record in caplog.records:
         assert record.levelname == "WARNING"
         assert "None of the input simulations meet the criteria selected" in record.message
+
+
+def test_simulate_filters_by_simulation_ids_if_the_list_has_atleast_one_valid_element(
+    tmp_path: pytest.TempPathFactory,
+):
+    # arrange
+    study = ps.ParametricStudy(tmp_path / "test_study")
+    material = AdditiveMaterial(name="test_material")
+    sb = SingleBeadInput(material=material)
+    p = PorosityInput(material=material)
+    ms = MicrostructureInput(material=material)
+    study.add_inputs([sb], priority=1)
+    study.add_inputs([p], priority=2)
+    study.add_inputs([ms], priority=3)
+    inputs = [p]
+    mock_additive = create_autospec(Additive)
+    mock_additive.material.return_value = material
+
+    # act
+    pr.simulate(study.data_frame(), mock_additive, simulation_ids=[p.id, "bogus"])
+
+    # assert
+    mock_additive.simulate.assert_called_once_with(inputs)
+
+
+@pytest.mark.parametrize(
+    "simulation_ids_input",
+    [
+        [],
+        None,
+    ],
+)
+def test_simulate_skips_filter_by_simulation_ids_if_the_list_is_empty_or_none(
+    simulation_ids_input, tmp_path: pytest.TempPathFactory
+):
+    # arrange
+    study = ps.ParametricStudy(tmp_path / "test_study")
+    material = AdditiveMaterial(name="test_material")
+    sb = SingleBeadInput(material=material)
+    p = PorosityInput(material=material)
+    ms = MicrostructureInput(material=material)
+    study.add_inputs([sb], priority=1)
+    study.add_inputs([p], priority=2)
+    study.add_inputs([ms], priority=3)
+    inputs = [sb, p, ms]
+    mock_additive = create_autospec(Additive)
+    mock_additive.material.return_value = material
+
+    # act
+    pr.simulate(study.data_frame(), mock_additive, simulation_ids=simulation_ids_input)
+
+    # assert
+    mock_additive.simulate.assert_called_once_with(inputs)
+
+
+def test_simulate_is_skipped_if_simulation_ids_list_has_invalid_elements(
+    tmp_path: pytest.TempPathFactory, caplog
+):
+    # arrange
+    study = ps.ParametricStudy(tmp_path / "test_study")
+    material = AdditiveMaterial(name="test_material")
+    sb = SingleBeadInput(material=material)
+    p = PorosityInput(material=material)
+    ms = MicrostructureInput(material=material)
+    study.add_inputs([sb], priority=1)
+    study.add_inputs([p], priority=2)
+    study.add_inputs([ms], priority=3)
+    mock_additive = create_autospec(Additive)
+    mock_additive.material.return_value = material
+
+    # act
+    result = pr.simulate(study.data_frame(), mock_additive, simulation_ids=["test_0", "test_4"])
+
+    # assert
+    mock_additive.simulate.assert_not_called()
+    assert result == []
+    assert len(caplog.records) == 3
+    assert "Simulation ID 'test_0' not found in the parametric study" in caplog.records[0].message
+    assert "Simulation ID 'test_4' not found in the parametric study" in caplog.records[1].message
+    assert "None of the input simulations meet the criteria selected" in caplog.records[2].message
+
+
+def test_simulate_filters_by_simulation_ids_and_skips_duplicates(tmp_path: pytest.TempPathFactory):
+    # arrange
+    study = ps.ParametricStudy(tmp_path / "test_study")
+    material = AdditiveMaterial(name="test_material")
+    sb = SingleBeadInput(material=material)
+    p = PorosityInput(material=material)
+    ms = MicrostructureInput(material=material)
+    study.add_inputs([sb], priority=1)
+    study.add_inputs([p], priority=2)
+    study.add_inputs([ms], priority=3)
+    inputs = [sb, p]
+    mock_additive = create_autospec(Additive)
+    mock_additive.material.return_value = material
+
+    # act
+    pr.simulate(study.data_frame(), mock_additive, simulation_ids=[sb.id, p.id, sb.id])
+
+    # assert
+    mock_additive.simulate.assert_called_once_with(inputs)
+
+
+def test_simulate_filters_by_simulation_ids_and_sorts_by_priority(tmp_path: pytest.TempPathFactory):
+    # arrange
+    study = ps.ParametricStudy(tmp_path / "test_study")
+    material = AdditiveMaterial(name="test_material")
+    sb = SingleBeadInput(material=material)
+    p = PorosityInput(material=material)
+    ms = MicrostructureInput(material=material)
+    study.add_inputs([sb], priority=1)
+    study.add_inputs([p], priority=2)
+    study.add_inputs([ms], priority=3)
+    inputs = [sb, ms]  # note that pr.simulate should reorder the inputs based on the priority
+    mock_additive = create_autospec(Additive)
+    mock_additive.material.return_value = material
+
+    # act
+    pr.simulate(study.data_frame(), mock_additive, simulation_ids=[ms.id, sb.id])
+
+    # assert
+    mock_additive.simulate.assert_called_once_with(inputs)
+
+
+def test_simulate_filters_by_simulation_ids_and_iteration(tmp_path: pytest.TempPathFactory):
+    # arrange
+    study = ps.ParametricStudy(tmp_path / "test_study")
+    material = AdditiveMaterial(name="test_material")
+    sb = SingleBeadInput(material=material)
+    p = PorosityInput(material=material)
+    ms = MicrostructureInput(material=material)
+    study.add_inputs([sb], iteration=1)
+    study.add_inputs([p], iteration=2)
+    study.add_inputs([ms], iteration=3)
+    inputs = [sb]
+    mock_additive = create_autospec(Additive)
+    mock_additive.material.return_value = material
+
+    # act
+    pr.simulate(
+        study.data_frame(),
+        mock_additive,
+        simulation_ids=[ms.id, p.id, sb.id],
+        iteration=1,
+    )
+
+    # assert
+    mock_additive.simulate.assert_called_once_with(inputs)
+
+
+def test_simulate_filters_by_simulation_ids_and_type(tmp_path: pytest.TempPathFactory):
+    # arrange
+    study = ps.ParametricStudy(tmp_path / "test_study")
+    material = AdditiveMaterial(name="test_material")
+    sb_1 = SingleBeadInput(bead_length=0.001, material=material)
+    sb_2 = SingleBeadInput(bead_length=0.002, material=material)
+    sb_3 = SingleBeadInput(bead_length=0.003, material=material)
+    sb_4 = SingleBeadInput(bead_length=0.004, material=material)
+    p = PorosityInput(material=material)
+    ms = MicrostructureInput(material=material)
+    study.add_inputs([sb_1, sb_2, sb_3, sb_4], iteration=1)
+    study.add_inputs([p], iteration=2)
+    study.add_inputs([ms], iteration=3)
+    inputs = [sb_1, sb_2]
+    mock_additive = create_autospec(Additive)
+    mock_additive.material.return_value = material
+
+    # act
+    pr.simulate(
+        study.data_frame(),
+        mock_additive,
+        simulation_ids=[sb_1.id, sb_2.id, "bogus"],
+        type=SimulationType.SINGLE_BEAD,
+    )
+
+    # assert
+    mock_additive.simulate.assert_called_once_with(inputs)
+
+
+def test_simulate_filters_by_simulation_ids_only_takes_pending_simulations(
+    tmp_path: pytest.TempPathFactory,
+):
+    # arrange
+    study = ps.ParametricStudy(tmp_path / "test_study")
+    material = AdditiveMaterial(name="test_material")
+    sb_1 = SingleBeadInput(bead_length=0.001, material=material)
+    sb_2 = SingleBeadInput(bead_length=0.002, material=material)
+    sb_3 = SingleBeadInput(bead_length=0.003, material=material)
+    sb_4 = SingleBeadInput(bead_length=0.004, material=material)
+    p = PorosityInput(material=material)
+    ms = MicrostructureInput(material=material)
+    melt_pool_msg = test_utils.get_test_melt_pool_message()
+    summary_1 = SingleBeadSummary(sb_1, melt_pool_msg, None)
+    summary_2 = SingleBeadSummary(sb_2, melt_pool_msg, None)
+    study.add_summaries([summary_1, summary_2])
+    study.add_inputs([sb_3, sb_4], iteration=1)
+    study.add_inputs([p], iteration=2)
+    study.add_inputs([ms], iteration=3)
+    inputs = [sb_3, sb_4, p, ms]
+    mock_additive = create_autospec(Additive)
+    mock_additive.material.return_value = material
+
+    # act
+    pr.simulate(
+        study.data_frame(),
+        mock_additive,
+        simulation_ids=[sb_1.id, sb_2.id, sb_3.id, sb_4.id, p.id, ms.id],
+    )
+
+    # assert
+    mock_additive.simulate.assert_called_once_with(inputs)